Moved RunningStat to utils.py; Added license information for the code from the OpenAI evolution-strategies repo
